compiling works, adding better error messages to generic factory

+++ include/genericfactory/LiteralStringList.h
@@ -8,82 +8,88 @@
 #include <string>
 
 //NOLINT(*)
-class literal_str_list {
- private:
-  const char* const text_ptr;
+class literal_str_list
+{
+private:
+  const char *const text_ptr;
   unsigned int text_size;
-  const literal_str_list* const head;
+  const literal_str_list *const head;
   constexpr char get_char_from_head(
-        unsigned int i,
-        unsigned int hd_size) const {
+      unsigned int i,
+      unsigned int hd_size) const
+  {
     return (i < hd_size ? (*head)[i]
-        : (i < (hd_size + text_size) ? text_ptr[i - hd_size] : '\0'));
+                        : (i < (hd_size + text_size) ? text_ptr[i - hd_size] : '\0'));
   }
   static constexpr std::size_t fnv_prime =
       (sizeof(std::size_t) == 8 ? 1099511628211u : 16777619u);
   static constexpr std::size_t fnv_offset =
       (sizeof(std::size_t) == 8 ? 14695981039346656037u : 2166136261u);
-  constexpr std::size_t fnv_1a_hash(unsigned int i) const {
-    return (i == 0 ?
-        (head != nullptr ?
-         ((head->fnv_1a_hash(head->text_size-1) ^ text_ptr[0]) * fnv_prime) :
-         fnv_offset) :
-        ((fnv_1a_hash(i-1) ^ text_ptr[i]) * fnv_prime));
+  constexpr std::size_t fnv_1a_hash(unsigned int i) const
+  {
+    return (i == 0 ? (head != nullptr ? ((head->fnv_1a_hash(head->text_size - 1) ^ text_ptr[0]) * fnv_prime) : fnv_offset) : ((fnv_1a_hash(i - 1) ^ text_ptr[i]) * fnv_prime));
   }
   template <typename FwdIter>
-  void copy_to_recurse(FwdIter& beg, FwdIter end) const {  // NOLINT
+  void copy_to_recurse(FwdIter &beg, FwdIter end) const
+  { // NOLINT
     if (head != nullptr)
       head->copy_to_recurse(beg, end);
     for (unsigned int i = 0; (i < text_size) && (beg != end); ++i, ++beg)
       *beg = text_ptr[i];
   }
-  void copy_to_recurse(char*& beg, char* end) const {  // NOLINT
+  void copy_to_recurse(char *&beg, char *end) const
+  { // NOLINT
     if (head != nullptr)
       head->copy_to_recurse(beg, end);
-    std::size_t sz_to_cpy = (end - beg < text_size ? end - beg : text_size);
+    std::size_t sz_to_cpy = (end - beg < static_cast<int>(text_size) ? end - beg : text_size);
     std::memcpy(beg, text_ptr, sz_to_cpy);
     beg += sz_to_cpy;
   }
-  constexpr literal_str_list(const char* aStr, unsigned int N,
-      const literal_str_list* aHead = nullptr) :
-    text_ptr(aStr), text_size(N), head(aHead) { }
+  constexpr literal_str_list(const char *aStr, unsigned int N,
+                             const literal_str_list *aHead = nullptr) : text_ptr(aStr), text_size(N), head(aHead) {}
 
- public:
+public:
   template <unsigned int N>
-  constexpr literal_str_list(const char(&aStr)[N],  // NOLINT
-      const literal_str_list* aHead = nullptr) :
-    text_ptr(aStr), text_size(N-1), head(aHead) {
-      static_assert(N >= 1, "Invalid string literal! Length is zero!");
+  constexpr literal_str_list(const char (&aStr)[N], // NOLINT
+                             const literal_str_list *aHead = nullptr) : text_ptr(aStr), text_size(N - 1), head(aHead)
+  {
+    static_assert(N >= 1, "Invalid string literal! Length is zero!");
   }
-  constexpr unsigned int size() const {
+  constexpr unsigned int size() const
+  {
     return text_size + (head != nullptr ? head->size() : 0);
   }
-  constexpr char operator[](unsigned int i) const {
-    return (head != nullptr ?
-        get_char_from_head(i, head->size()) :
-        (i < text_size ? text_ptr[i] : '\0'));
+  constexpr char operator[](unsigned int i) const
+  {
+    return (head != nullptr ? get_char_from_head(i, head->size()) : (i < text_size ? text_ptr[i] : '\0'));
   }
   template <unsigned int N>
-  constexpr literal_str_list operator+(const char(&aHead)[N]) const {  // NOLINT
+  constexpr literal_str_list operator+(const char (&aHead)[N]) const
+  { // NOLINT
     return literal_str_list(aHead, this);
   }
-  constexpr literal_str_list operator+(const literal_str_list& aHead) const {
+  constexpr literal_str_list operator+(const literal_str_list &aHead) const
+  {
     return literal_str_list(aHead.text_ptr, aHead.text_size, this);
   }
-  constexpr std::size_t hash() const {
-    return fnv_1a_hash(text_size-1);
+  constexpr std::size_t hash() const
+  {
+    return fnv_1a_hash(text_size - 1);
   }
   template <typename FwdIter>
-  void copy_to(FwdIter beg, FwdIter end) const {
+  void copy_to(FwdIter beg, FwdIter end) const
+  {
     copy_to_recurse(beg, end);
   }
-  void copy_to(char* beg, char* end) const {
+  void copy_to(char *beg, char *end) const
+  {
     copy_to_recurse(beg, end);
   }
 };
-inline std::string convert_to_string(const literal_str_list& lit) {
+inline std::string convert_to_string(const literal_str_list &lit)
+{
   std::string result(lit.size(), ' ');
   lit.copy_to(result.begin(), result.end());
   return result;
 }
-#endif  // GENERICFACTORY_LITERALSTRINGLIST_H_
+#endif // GENERICFACTORY_LITERALSTRINGLIST_H_
